AuthContext.js to tsx
